refactor: moved utils file into utils directory
