+++ src/lib/utils/utils.ts
@@ -1,6 +1,6 @@
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
-import { AuthState, Footwear } from "./types";
+import { AuthState, Footwear } from "../types";
 import { redirect } from "@tanstack/react-router";
 
 export function cn(...inputs: ClassValue[]) {
